Created new golfer homepage, renamed old component for now
